tests: Update config tests
